Split special infected features into optional include
